requirements.txt
absl-py==1.2.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.6.0
black==22.6.0
bleach==5.0.1
cachetools==5.2.0
celluloid==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chex==0.1.3
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
coverage==5.5
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.7
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.4.1
entrypoints==0.4
etils==0.6.0
executing==0.9.0
fairseq==0.12.2
fastjsonschema==2.16.1
filelock==3.8.0
flatbuffers==22.10.26
flax==0.5.2
fonttools==4.34.4
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.2
gast==0.4.0
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.10.1
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jax==0.3.15
jaxlib==0.3.15+cuda11.cudnn82
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
llvmlite==0.39.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mne==1.0.3
mne-bids==0.10
mnelab==0.8.4
moabb==0.4.6
msgpack==1.0.4
multidict==6.0.5
mypy-extensions==0.4.3
natsort==8.2.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.2.1
notebook==6.4.12
numba==0.56.0
numpy==1.22.4
oauthlib==3.2.0
omegaconf==2.2.3
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.2
pooch==1.6.0
portalocker==2.5.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
pymatreader==0.0.30
pynndescent==0.5.7
pyparsing==3.0.9
pyriemann==0.3
pyrsistent==0.18.1
PySide6==6.2.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyxdf==1.16.3
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.9
sacrebleu==2.2.1
scikit-learn==1.1.1
scikit-network==0.30.0
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.9.5
setproctitle==1.3.2
shiboken6==6.2.4
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu113
torch_geometric==2.5.2
torchaudio==0.12.1+cu113
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.24.0
typing_extensions==4.3.0
umap==0.1.1
umap-learn==0.5.3
urllib3==1.26.11
wandb==0.13.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.0
widgetsnbextension==3.6.1
wrapt==1.14.1
xmltodict==0.13.0
yapf==0.32.0
yarg==0.1.9
yarl==1.9.4
zipp==3.8.1
